added features

+++ frontend/src/app/classes/CommonList.ts
@@ -1,6 +1,7 @@
-import { OnInit } from '@angular/core';
+import { OnInit, ViewChild } from '@angular/core';
 import { DataTableHelper } from './DataTableHelper';
 import { ServerInfo } from './ServerInfo';
+import { DataTableDirective } from 'angular-datatables';
 
 export class CommonList implements OnInit
 {
@@ -12,30 +13,51 @@ export class CommonList implements OnInit
     dataTableUrl = "/api/"+this.url+"-datatable";
     tableColumns: any;
 
-    constructor(_http, _router, _url, _tableColumns) {
+    @ViewChild(DataTableDirective, {static: false})
+    dtElement: DataTableDirective;
+    currentUser: any;
+    auth: any;
+
+    constructor(_http, _router, _url, _tableColumns, _auth) {
         this.http = _http;
         this.router = _router;       
         this.url = _url;
         this.dataTableUrl = "/api/"+this.url+"-datatable";
         this.tableColumns = _tableColumns;
+        this.auth = _auth;
     }
 
-    ngOnInit(): void {
+    async ngOnInit() {
         this.dtOptions = DataTableHelper.generateDataTableOptions(this, this.dataTableUrl, this.tableColumns);
+        
+        this.currentUser = this.auth.currentUser;
+
+        if(this.currentUser == null) {
+            this.currentUser = await this.auth.getCurrentUser();
+        }
     }
 
     edit(id) {
         this.router.navigate([this.url+ '/' + id]);
     }
 
+    rerender_datatable() {
+        this.dtElement.dtInstance.then((dtInstance: DataTables.Api) => {
+            dtInstance.draw();
+        });
+    }
+
     delete(id) {
+
+        if( confirm("Are you sure to delete?") == false ) {
+            return;
+        }
+
         this.http.delete(ServerInfo.Url + "/api/"+this.url+"/" + id)
         .subscribe(data => {
+            alert(data['message']);
             if(data['success']) {
-                alert(data['message']);
-                this.data = (<Array<any>>this.data).filter(item => {
-                item.id != id;
-                });
+                this.rerender_datatable();
             }
         })
     }
